requirements.txt
werkzeug
flask
boto3
python-dotenv
numpy-stl
matplotlib
pymongo
